requirements.txt
black
Jinja2
mkdocs
pytest
pytest-cov
pytest-env
pytest-mock
PyYAML
click
setuptools-git-versioning
twine


